tg_interface/requirements.txt
APScheduler==3.6.3
black==22.6.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
configparser==5.2.0
flake8==4.0.1
idna==3.3
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tomli==2.0.1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
